requirements.txt
Pillow
pandas
qrcode
opencv-python
pywebview